+++ CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.8.0 - 2023-01-04
+
+### Added
+* Store action toasters
+* Change sequence of operations
+
+
 ## 0.7.6 - 2022-12-27
 
 ### Added